auth/Crypto: fix TAINTED_SCALAR
Make sure we pass only size_t to get_random_bytes() as size parameter
instead of int to prevent usage of a tainted variable when e.g. a
negative value is passed to read() for nbyte.

Fix for:
CID 1219616 (#1 of 1): Use of untrusted scalar value (TAINTED_SCALAR)
 tainted_data: Passing tainted variable buf to a tainted sink.
 tainted_data_argument: Calling function cls_gen_random_bytes taints argument buf.
CID 1219617 (#1 of 1): Use of untrusted scalar value (TAINTED_SCALAR)
 tainted_data_argument: Calling function get_random_bytes taints argument buf.
 tainted_data: Passing tainted variable buf to a tainted sink.

+++ src/auth/Crypto.cc
@@ -36,7 +36,7 @@
 
 #include <errno.h>
 
-int get_random_bytes(char *buf, int len)
+int get_random_bytes(char *buf, size_t len)
 {
   int fd = TEMP_FAILURE_RETRY(::open("/dev/urandom", O_RDONLY));
   if (fd < 0)
@@ -46,7 +46,7 @@ int get_random_bytes(char *buf, int len)
   return ret;
 }
 
-static int get_random_bytes(int len, bufferlist& bl)
+static int get_random_bytes(size_t len, bufferlist& bl)
 {
   char buf[len];
   get_random_bytes(buf, len);
